requirements.txt
Flask
Tensorflow-gpu
opencv-python
nltk